+++ include/input/analog.h
@@ -24,6 +24,6 @@ void analog_init();
 void analog_access_safe(analog_data_s *out, analog_access_t type);
 
 void analog_config_command(analog_cmd_t cmd, webreport_cmd_confirm_t cb);
-void analog_task(uint32_t timestamp);
+void analog_task(uint64_t timestamp);
 
 #endif
