+++ include/usb/gcinput.h
@@ -59,6 +59,6 @@ extern const uint8_t gc_hid_report_descriptor[];
 extern const uint8_t gc_configuration_descriptor[];
 
 void gcinput_enable(bool enable);
-void gcinput_hid_report(uint32_t timestamp, hid_report_tunnel_cb cb);
+void gcinput_hid_report(uint64_t timestamp, hid_report_tunnel_cb cb);
 
 #endif
