+++ include/usb/webusb.h
@@ -22,6 +22,6 @@ bool webusb_outputting_check();
 void webusb_command_confirm_cb(cfg_block_t config_block, uint8_t cmd, bool success, uint8_t *data, uint32_t size);
 void webusb_send_bulk(const uint8_t *data, uint16_t size);
 void webusb_command_handler(uint8_t *data, uint32_t size);
-void webusb_send_rawinput(uint32_t timestamp);
+void webusb_send_rawinput(uint64_t timestamp);
 
 #endif
