+++ include/utilities/interval.h
@@ -10,12 +10,12 @@ extern "C" {
 
 typedef struct 
 {
-  uint32_t this_time;
-  uint32_t last_time;
+  uint64_t this_time;
+  uint64_t last_time;
 } interval_s;
 
-bool interval_run(uint32_t timestamp, uint32_t interval, interval_s *state);
-bool interval_resettable_run(uint32_t timestamp, uint32_t interval, bool reset, interval_s *state);
+bool interval_run(uint64_t timestamp, uint32_t interval, interval_s *state);
+bool interval_resettable_run(uint64_t timestamp, uint32_t interval, bool reset, interval_s *state);
 
 #ifdef __cplusplus
 }
